+++ src/VoltManager.php
@@ -89,7 +89,7 @@ class VoltManager
     /**
      * Set the currently logged in user for the application.
      */
-    public function actingAs(Authenticatable $user, string $driver = null): static
+    public function actingAs(Authenticatable $user, ?string $driver = null): static
     {
         $this->manager->actingAs($user, $driver);